Refactor ArtistServiceCheckboxCard and remove ArtistServiceProduct component
