--- app/components/artist/ArtistServiceProduct.module.css
@@ -1,27 +1,0 @@
-.card {
-  transition: transform 150ms ease, box-shadow 150ms ease;
-
-  &:hover {
-    transform: scale(1.01);
-    box-shadow: var(--mantine-shadow-md);
-  }
-}
-
-.image {
-  display: none;
-  @media (min-width: $mantine-breakpoint-sm) {
-    display: block;
-    flex: 0 0 20%;
-  }
-}
-
-.title {
-  font-family: 'Greycliff CF', var(--mantine-font-family);
-  font-weight: bold;
-}
-
-.unset {
-  border-bottom: var(--_card-section-border-bottom, unset);
-  border-top: var(--_card-section-border-top, unset);
-  display: block;
-}
