+++ app/routes/artist_.$username.treatment.$productHandle.pick-more.tsx
@@ -1,4 +1,17 @@
-import {Button, SimpleGrid, Skeleton, Text, Title} from '@mantine/core';
+import {
+  Badge,
+  Box,
+  Button,
+  Divider,
+  Flex,
+  Group,
+  Image,
+  SimpleGrid,
+  Skeleton,
+  Text,
+  Title,
+  rem,
+} from '@mantine/core';
 import {Await, Link, useLoaderData, useSearchParams} from '@remix-run/react';
 import {defer, type LoaderFunctionArgs} from '@shopify/remix-oxygen';
 import {Suspense} from 'react';
@@ -7,9 +20,12 @@ import {getBookingShopifyApi} from '~/lib/api/bookingShopifyApi';
 import {type CustomerProductBase} from '~/lib/api/model';
 import {ALL_PRODUCTS_QUERY} from './artist.$username._index';
 
-import {ArtistServiceProduct} from '~/components/artist/ArtistServiceProduct';
+import {Money, Image as ShopifyImage, parseGid} from '@shopify/hydrogen';
+
+import {ArtistServiceCheckboxCard} from '~/components/artist/ArtistServiceCheckboxCard';
 import {ArtistShell} from '~/components/ArtistShell';
 import {TreatmentStepper} from '~/components/TreatmentStepper';
+import {durationToTime} from '~/lib/duration';
 
 export function shouldRevalidate() {
   return false;
@@ -131,5 +147,104 @@ function RenderArtistProducts({products, services}: RenderArtistProductsProps) {
         {restProductsMarkup}
       </SimpleGrid>
     </>
+  );
+}
+
+type ArtistServiceProductProps = {
+  product: ProductItemFragment;
+  services: CustomerProductBase[];
+  defaultChecked?: boolean;
+};
+
+function ArtistServiceProduct({
+  product,
+  services,
+  defaultChecked,
+}: ArtistServiceProductProps) {
+  const [, setSearchParams] = useSearchParams();
+
+  const onChange = (checked: boolean) => {
+    setSearchParams((prev) => {
+      const existingItems = prev.getAll('productIds');
+      const itemIndex = existingItems.indexOf(parseGid(product.id).id);
+
+      if (checked && itemIndex === -1) {
+        prev.append('productIds', parseGid(product.id).id);
+      }
+
+      if (!checked) {
+        const updatedItems = existingItems.filter(
+          (item) => item !== parseGid(product.id).id,
+        );
+        prev.delete('productIds');
+        updatedItems.forEach((item) => prev.append('productIds', item));
+      }
+      return prev;
+    });
+  };
+
+  const artistService = services.find(({productId}) => {
+    return productId.toString() === parseGid(product.id).id;
+  });
+
+  const leftSection = (
+    <Text c="dimmed" size="xs" tt="uppercase" fw={700}>
+      {durationToTime(artistService?.duration ?? 0)}
+    </Text>
+  );
+
+  const rightSection = artistService?.price && (
+    <Badge variant="light" color="gray" size="md">
+      <Money data={artistService?.price as any} />
+    </Badge>
+  );
+
+  return (
+    <ArtistServiceCheckboxCard
+      value={artistService!.productId.toString()}
+      defaultChecked={defaultChecked}
+      onChange={onChange}
+      name="productIds"
+    >
+      <Flex>
+        {product.featuredImage && (
+          <Image
+            component={ShopifyImage}
+            data={product.featuredImage}
+            h="auto"
+            loading="lazy"
+            visibleFrom="sm"
+          />
+        )}
+        <Flex direction="column" style={{flex: '1'}}>
+          <Box p="xs" style={{flex: 1}}>
+            <Title order={4} mb={rem(4)} fw={500}>
+              {product.title}
+            </Title>
+            <Flex
+              gap="sm"
+              direction="column"
+              justify="flex-start"
+              style={{flexGrow: 1, position: 'relative'}}
+            >
+              <Text c="dimmed" size="sm" fw={400} lineClamp={3}>
+                {artistService?.description ||
+                  product.description ||
+                  'ingen beskrivelse'}
+              </Text>
+            </Flex>
+          </Box>
+
+          <Divider />
+
+          <Box p="xs">
+            <Group justify="space-between">
+              {leftSection}
+              {rightSection}
+            </Group>
+          </Box>
+        </Flex>
+      </Flex>
+    </ArtistServiceCheckboxCard>
   );
 }
